requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
redis
python-jose[cryptography]
passlib
python-dotenv
httpx
python-multipart
